fix: twMergeConfigの副作用がランタイムで実行されない問題を修正
