+++ packages/smarthr-ui/package.json
@@ -170,7 +170,8 @@
     "scaffold:storybook": "plop"
   },
   "sideEffects": [
-    "*.css"
+    "*.css",
+    "**/configureTwMerge.*"
   ],
   "typings": "lib/index.d.ts"
 }
